Moving tree_shell_node_adder.py under tree module
